requirements.txt
kafka-python==2.0.2
numpy==1.24.1
py4j==0.10.9
pyarrow==10.0.1
pyspark==3.0.0
PyYAML==6.0
